+++ src/cluster_mgr/src/cli/task/runtime_deps_install.rs
@@ -101,7 +101,7 @@ impl TaskExecutor for RuntimeDepsInstallation {
         ssh_session.close().await?;
         task_return_value!(
             install_dep_cmd_rs,
-            |status_code: usize| -> CmdErr {
+            |status_code: i32| -> CmdErr {
                 CmdErr::ExecUserCmdErr(self.install_dep_cmd.clone(), status_code.to_string())
             },
             "RuntimeDepsInstallation"
